Split app-state logic from view rendering (step 5 of layered refactor)
Each of the 13 app/<screen>.rs modules interleaved event-handling/state
logic with rect-geometry/draw-command building. Split every screen into
core/state/<screen>.rs (handle_*_event, open_*, decisions — no Rect, no
Painter, no draw calls) and ui/view/<screen>.rs (the *_rect geometry
functions and render_* draw calls), keeping them as impl App blocks so
the split is a pure move with no behavior change.

Added core/effect.rs with an Effect enum and an execute() that runs it;
Diagnostics' show-logs toggle is wired through it end to end as the
first case (handle_diagnostics_event now returns Vec<Effect>, and
main.rs's dispatch executes them) — establishing the pattern for the
rest, deferred (see docs/HANDOFF_layered_refactor.md).

App itself, and mod.rs's cross-screen plumbing (draw_list, prepare_tiles,
mouse hit-testing, tick_animations), stay in app/mod.rs: App holds
tiny_skia Painter tile caches, so moving it into core/state would fail
the plan's own "no tiny_skia in core" check without first splitting the
view-tile cache out of App — a larger, separate piece of work.
