+++ src/core/state/diagnostics.rs
@@ -1,8 +1,11 @@
-use super::*;
-use crate::ui::render::Rect;
+//! Diagnostics screen logic. Rendering lives in `ui::view::diagnostics`.
+use crate::app::App;
+use crate::app::DropdownState;
+use crate::core::effect::Effect;
+use crate::core::screen::Screen;
+use crate::services::store;
+use crate::ui::{self, MenuEvent};
 use std::time::Instant;
-
-use crate::ui::{self, FocusRow, MenuEvent, Painter};
 
 impl App {
     /// Opens the Diagnostics screen — reached from the "Diagnostics" row at the
@@ -14,23 +17,14 @@ impl App {
         self.screen = Screen::Diagnostics;
     }
 
-    pub(crate) fn diagnostics_rows(&self) -> Vec<FocusRow> {
-        ui::diagnostics_rows(&self.settings)
-    }
-
-    pub(crate) fn diagnostics_subtitle(&self) -> String {
-        "Debug aids for on-device investigation.".to_string()
-    }
-
-    pub(crate) fn diagnostics_card_rect(screen_w: u32, screen_h: u32, fonts: &ui::Fonts, subtitle: &str) -> Rect {
-        ui::list_modal_card_rect(screen_w, screen_h, fonts, subtitle, ui::DIAGNOSTICS_ROW_COUNT)
-    }
-
     /// `ui::DIAG_ROW_*` rows: Log level opens the same dropdown picker every
     /// `Settings` dropdown uses (its row `0` is disambiguated from `Settings`' row 0
     /// by `self.screen`, see `dropdown_overlay_tile`'s docs); the rest are plain
     /// Left/Right/Confirm toggles. Back saves and returns to Settings.
-    pub(crate) fn handle_diagnostics_event(&mut self, ev: MenuEvent) {
+    ///
+    /// Returns effects the runtime should execute (currently just the log-overlay
+    /// toggle — see `core::effect` for scope/rationale).
+    pub(crate) fn handle_diagnostics_event(&mut self, ev: MenuEvent) -> Vec<Effect> {
         if let Some(dd) = self.dropdown.as_mut() {
             let len = ui::LOG_LEVEL_OPTIONS.len();
             match ev {
@@ -48,13 +42,14 @@ impl App {
                 }
                 MenuEvent::Left | MenuEvent::Right | MenuEvent::Secondary => {}
             }
-            return;
+            return Vec::new();
         }
         let len = self.diagnostics_rows().len();
         if ui::list_nav(&mut self.diagnostics_focused, len, ev) {
             self.modal_focus_anim = Some(Instant::now());
-            return;
+            return Vec::new();
         }
+        let mut effects = Vec::new();
         match (self.diagnostics_focused, ev) {
             (ui::DIAG_ROW_LOG_LEVEL, MenuEvent::Left | MenuEvent::Right) => self.cycle_log_level(),
             (ui::DIAG_ROW_LOG_LEVEL, MenuEvent::Confirm) => {
@@ -72,7 +67,7 @@ impl App {
             (ui::DIAG_ROW_SHOW_LOGS, MenuEvent::Left | MenuEvent::Right | MenuEvent::Confirm) => {
                 let from = self.settings.show_logs;
                 self.settings.show_logs = !from;
-                crate::real::set_log_overlay_enabled(!from);
+                effects.push(Effect::SetLogOverlay(!from));
                 self.switch_anim = Some((Instant::now(), from));
             }
             (ui::DIAG_ROW_SEND_LOGS, MenuEvent::Confirm) => {
@@ -88,6 +83,7 @@ impl App {
             }
             _ => {}
         }
+        effects
     }
 
     fn set_log_level(&mut self, level: store::LogLevelOverride) {
@@ -99,27 +95,5 @@ impl App {
         let idx = ui::log_level_dropdown_current_index(self.settings.log_level_override);
         let next = ui::cycle_index(idx, ui::LOG_LEVEL_OPTIONS.len(), true);
         self.set_log_level(ui::LOG_LEVEL_OPTIONS[next]);
-    }
-
-    pub(crate) fn render_diagnostics(
-        &self,
-        painter: &mut Painter,
-        text_cache: &mut crate::ui::TextCache,
-        fonts: &ui::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-    ) -> Result<()> {
-        let subtitle = self.diagnostics_subtitle();
-        let card = Self::diagnostics_card_rect(screen_w, screen_h, fonts, &subtitle);
-        self.draw_modal_shell(painter, text_cache, fonts.raster, fonts.icon, card)?;
-        ui::render_list_modal(
-            painter,
-            text_cache,
-            fonts,
-            card,
-            "Diagnostics",
-            &subtitle,
-            &self.diagnostics_rows(),
-        )
     }
 }
