+++ src/core/state/experimental.rs
@@ -1,8 +1,8 @@
-use super::*;
-use crate::ui::render::Rect;
+//! Experimental screen logic. Rendering lives in `ui::view::experimental`.
+use crate::app::App;
+use crate::core::screen::Screen;
+use crate::ui::{self, MenuEvent};
 use std::time::Instant;
-
-use crate::ui::{self, FocusRow, MenuEvent, Painter};
 
 impl App {
     /// Opens the Experimental screen (Settings → `ui::ROW_EXPERIMENTAL`). Holds unstable,
@@ -12,18 +12,6 @@ impl App {
         // Stash scroll so Back can restore it; Experimental doesn't use it.
         self.settings_scroll = self.scroll;
         self.screen = Screen::Experimental;
-    }
-
-    pub(crate) fn experimental_rows(&self) -> Vec<FocusRow> {
-        ui::experimental_rows(&self.settings)
-    }
-
-    pub(crate) fn experimental_subtitle(&self) -> String {
-        "Unstable, off by default. Frame pacer also toggles live with the Blue button.".to_string()
-    }
-
-    pub(crate) fn experimental_card_rect(screen_w: u32, screen_h: u32, fonts: &ui::Fonts, subtitle: &str) -> Rect {
-        ui::list_modal_card_rect(screen_w, screen_h, fonts, subtitle, ui::EXPERIMENTAL_ROW_COUNT)
     }
 
     /// All rows are plain Left/Right/Confirm toggles. Back saves and returns to Settings.
@@ -46,27 +34,5 @@ impl App {
             }
             _ => {}
         }
-    }
-
-    pub(crate) fn render_experimental(
-        &self,
-        painter: &mut Painter,
-        text_cache: &mut crate::ui::TextCache,
-        fonts: &ui::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-    ) -> Result<()> {
-        let subtitle = self.experimental_subtitle();
-        let card = Self::experimental_card_rect(screen_w, screen_h, fonts, &subtitle);
-        self.draw_modal_shell(painter, text_cache, fonts.raster, fonts.icon, card)?;
-        ui::render_list_modal(
-            painter,
-            text_cache,
-            fonts,
-            card,
-            "Experimental",
-            &subtitle,
-            &self.experimental_rows(),
-        )
     }
 }
