+++ src/core/state/hostmenu.rs
@@ -1,16 +1,13 @@
-//! The per-host actions menu — the extension point for anything that acts on one host.
-//!
-//! Split out of the former single-file `app.rs`; see `super`'s module docs.
+//! The per-host actions menu — logic. Rendering lives in `ui::view::hostmenu`.
 //!
 //! Adding an action here is deliberately two edits and nothing else: a row in
 //! [`App::host_menu_actions`] and an arm in [`App::confirm_host_menu_row`]. Everything
 //! else — card geometry, the unfocused shell, the focused-row tile, the focus pop — is
 //! `ui::ListModal`'s, shared with any future list screen.
-use super::*;
-use crate::ui::render::Rect;
+use crate::app::App;
+use crate::core::screen::Screen;
+use crate::ui::{FocusRow, HostEntry, MenuEvent};
 use std::time::Instant;
-
-use crate::ui::{self, FocusRow, HostEntry, MenuEvent, Painter};
 
 /// Host action (enum instead of bare index so conditional rows don't silently shift indices).
 #[derive(Clone, Copy, PartialEq, Eq)]
@@ -49,17 +46,20 @@ impl App {
             (
                 HostAction::Connect,
                 if entry.is_paired() {
-                    FocusRow::action(ui::ICON_TV, "Connect")
+                    FocusRow::action(crate::ui::ICON_TV, "Connect")
                 } else {
                     // The hint goes in the value column like every other Action row's,
                     // rather than being parenthesised into the label.
-                    FocusRow::action_with_value(ui::ICON_TV, "Connect", "pairs first")
+                    FocusRow::action_with_value(crate::ui::ICON_TV, "Connect", "pairs first")
                 },
             ),
-            (HostAction::Pair, FocusRow::action(ui::ICON_LOCK, "Pair with PIN…")),
+            (
+                HostAction::Pair,
+                FocusRow::action(crate::ui::ICON_LOCK, "Pair with PIN…"),
+            ),
             (
                 HostAction::SpeedTest,
-                FocusRow::action(ui::ICON_SIGNAL, "Test network speed…"),
+                FocusRow::action(crate::ui::ICON_SIGNAL, "Test network speed…"),
             ),
         ];
         if !entry.mac().is_empty() {
@@ -71,21 +71,20 @@ impl App {
             // underneath can't bake in a highlight that outlives it.
             rows.push((
                 HostAction::Wake,
-                FocusRow::action(ui::ICON_POWER, "Wake host").with_menu(false),
+                FocusRow::action(crate::ui::ICON_POWER, "Wake host").with_menu(false),
             ));
         }
         if saved {
-            rows.push((HostAction::Edit, FocusRow::action(ui::ICON_EDIT, "Edit address…")));
+            rows.push((
+                HostAction::Edit,
+                FocusRow::action(crate::ui::ICON_EDIT, "Edit address…"),
+            ));
             rows.push((
                 HostAction::Forget,
-                FocusRow::action(ui::ICON_DELETE, "Forget host").danger(),
+                FocusRow::action(crate::ui::ICON_DELETE, "Forget host").danger(),
             ));
         }
         rows
-    }
-
-    pub(crate) fn host_menu_rows(&self) -> Vec<FocusRow> {
-        self.host_menu_actions().into_iter().map(|(_, r)| r).collect()
     }
 
     /// The host's name — the menu's title.
@@ -109,20 +108,10 @@ impl App {
             })
     }
 
-    pub(crate) fn host_menu_card_rect(
-        screen_w: u32,
-        screen_h: u32,
-        fonts: &ui::Fonts,
-        subtitle: &str,
-        rows: usize,
-    ) -> Rect {
-        ui::list_modal_card_rect(screen_w, screen_h, fonts, subtitle, rows)
-    }
-
     /// Handles host menu events; may return `ConnectTarget` (currently never does).
     pub(crate) fn handle_host_menu_event(&mut self, ev: MenuEvent) {
         let len = self.host_menu_actions().len();
-        if ui::list_nav(&mut self.menu_focused, len, ev) {
+        if crate::ui::list_nav(&mut self.menu_focused, len, ev) {
             // Vertical movement always lands on the row body — a ⋯ belongs to the row
             // it's on, so leaving that row leaves the button too.
             self.host_menu_dots = false;
@@ -182,28 +171,5 @@ impl App {
             HostAction::Edit => self.open_edit_host(idx),
             HostAction::Forget => self.open_forget_host(idx),
         }
-    }
-
-    pub(crate) fn render_host_menu(
-        &self,
-        painter: &mut Painter,
-        text_cache: &mut crate::ui::TextCache,
-        fonts: &ui::Fonts,
-        screen_w: u32,
-        screen_h: u32,
-    ) -> Result<()> {
-        let rows = self.host_menu_rows();
-        let subtitle = self.host_menu_subtitle();
-        let card = Self::host_menu_card_rect(screen_w, screen_h, fonts, &subtitle, rows.len());
-        self.draw_modal_shell(painter, text_cache, fonts.raster, fonts.icon, card)?;
-        ui::render_list_modal(
-            painter,
-            text_cache,
-            fonts,
-            card,
-            &self.host_menu_title(),
-            &subtitle,
-            &rows,
-        )
     }
 }
