+++ src/platform/webos/glibc_compat_shim.c
